moved page and librarymanager.py
