+++ packages/algomancy-content/src/algomancy_content/__init__.py
@@ -19,7 +19,7 @@ from .pages import (
     PlaceholderScenarioPage,
     StandardOverviewPage,
 )
-from .librarymanager import LibraryManager
+from algomancy_gui.librarymanager import LibraryManager
 
 __all__ = [
     "LibraryManager",
