+++ packages/algomancy-content/src/algomancy_content/pages/standardoverviewpage.py
@@ -3,7 +3,7 @@ from typing import Any
 from dash import html, dash_table
 
 from algomancy_scenario import Scenario
-from .page import BaseOverviewPage
+from algomancy_gui.page import BaseOverviewPage
 
 OVERVIEW_TABLE = "overview-table"
