+++ packages/algomancy-gui/src/algomancy_gui/contentregistry.py
@@ -3,7 +3,7 @@ from typing import Callable, List
 from dash import html
 
 from algomancy_data import BASE_DATA_BOUND
-from algomancy_content.pages.page import (
+from algomancy_gui.page import (
     BaseHomePage,
     BaseDataPage,
     BaseScenarioPage,
