+++ packages/algomancy-gui/src/algomancy_gui/page.py
